requirements.txt
# PostgreSQL драйвер psycopg 3 (с бинарной сборкой)
psycopg[binary]==3.2.1

# SQL инструменты
sqlalchemy==2.0.32
alembic==1.13.2

# Базовые зависимости
python-dotenv==1.0.1
requests==2.32.3
vk-api==11.10.0
python-telegram-bot==21.3
python-dateutil==2.9.0.post0
pydantic==2.8.2
loguru==0.7.2
colorama==0.4.6
tqdm==4.66.5

# Для работы с датами
pytz==2024.1